+++ region/views/mining.py
@@ -67,7 +67,7 @@ def mining(request):
             bonus = raw_top[0][1] // 10
 
     if bonus > 0:
-        daily_limit = daily_limit * (1 + (bonus / 100))
+        daily_limit = int(daily_limit * (1 + (bonus / 100)))
 
     # ------------------
